Support float segment_sec and frame_interval_sec=0 for every-frame extraction
- segment_sec now accepts float values (e.g., 0.5 for half-second clips)
- frame_interval_sec=0 extracts every frame instead of crashing

+++ pkg/models/manifest.go
@@ -22,7 +22,7 @@ type VideoInputInfo struct {
 }
 
 type ProcessingParams struct {
-	SegmentSec       int    `json:"segment_sec"`
+	SegmentSec       float64 `json:"segment_sec"`
 	FrameIntervalSec int    `json:"frame_interval_sec"`
 	FrameFormat      string `json:"frame_format"`
 	ZipFormat        string `json:"zip_format"`
